mac edited, fixed override syntax

+++ shapes/shapetrnsp.cpp
@@ -87,7 +87,7 @@ void ShapeTrnsp::Parse(QString& src)
     roundedTuRect = QRectF(XY + QPointF(-3, -3), QSizeF(width+6, height + 6));
 
     // ищем описатель свойств и рассчитываем графику с учетом координат транспаранта
-    prop = idObj>=0 && idObj<TrnspDescription::descriptions.size() ? TrnspDescription::descriptions[idObj] : nullptr;
+    prop = idObj>=0 && idObj<(int)TrnspDescription::descriptions.size() ? TrnspDescription::descriptions[idObj] : nullptr;
     if (prop == nullptr)
         set->logger()->log(QString("ShapeTrnsp. Не найдено описание транспаранта с типом #%1: '%2'").arg(idObj).arg(src));
     else
